finish backend

+++ apps/backend/src/endpoints/task/TaskService.ts
@@ -1,15 +1,34 @@
-import { Prisma } from "@prisma/client";
+import { DecodedIdToken } from "firebase-admin/auth";
+import { API } from "packages/api";
 import HttpException from "../../utils/HttpException";
 import { prisma } from "../../utils/Prisma";
 
-export async function findAllTasksByCompanyId(companyId: string) {
+export async function findAllTasksByCompanyId(
+	companyId: string,
+	requestingUser: DecodedIdToken
+): Promise<API.CompletedTask[]> {
+	let company;
 	try {
-		await prisma.company.findUnique({
+		company = await prisma.company.findUnique({
 			where: {
 				id: companyId,
 			},
+			include: {
+				users: true,
+			},
 		});
-		const foundTasks = await prisma.completedTask.findMany({
+	} catch (e) {
+		throw new HttpException(404, "Company not found");
+	}
+
+	if (company.users.find(user => user.userId === requestingUser.uid)) {
+		throw new HttpException(403, "You are not allowed to view this company");
+	}
+
+	let foundTasks: API.CompletedTask[] = [];
+
+	try {
+		foundTasks = await prisma.completedTask.findMany({
 			where: {
 				companyId: companyId,
 			},
@@ -17,57 +36,75 @@ export async function findAllTasksByCompanyId(companyId: string) {
 				taskId: true,
 			},
 		});
-		if (Object.keys(foundTasks).length == 0) {
-			throw new HttpException(404, "No completed tasks found in this company");
-		}
-		return foundTasks;
-	} catch (e) {
-		throw new HttpException(e.status || 500, e.message);
-	}
+	} catch (e) {}
+
+	return foundTasks;
 }
 
-export async function addTaskToCompany(companyId: string, taskId: string) {
+export async function addTaskToCompany(
+	companyId: string,
+	taskId: string,
+	requestingUser: DecodedIdToken
+) {
+	let company;
 	try {
-		await prisma.company.findUnique({
+		company = await prisma.company.findUnique({
 			where: {
 				id: companyId,
 			},
+			include: {
+				users: true,
+			},
 		});
-		const createdTask = await prisma.completedTask.create({
+	} catch (e) {
+		throw new HttpException(404, "Company not found");
+	}
+
+	if (company.users.find(user => user.userId === requestingUser.uid)) {
+		throw new HttpException(403, "You are not allowed to view this company");
+	}
+
+	try {
+		await prisma.completedTask.create({
 			data: {
 				companyId: companyId,
 				taskId: taskId,
 			},
 		});
-		return createdTask;
 	} catch (e) {
 		throw new HttpException(500, e.message);
 	}
 }
 
-export async function deleteTaskFromCompany(companyId: string, taskId: string) {
+export async function deleteTaskFromCompany(
+	companyId: string,
+	taskId: string,
+	requestingUser: DecodedIdToken
+) {
+	let company;
 	try {
-		await prisma.company.findUnique({
+		company = await prisma.company.findUnique({
 			where: {
 				id: companyId,
 			},
+			include: {
+				users: true,
+			},
 		});
+	} catch (e) {
+		throw new HttpException(404, "Company not found");
+	}
+
+	if (company.users.find(user => user.userId === requestingUser.uid)) {
+		throw new HttpException(403, "You are not allowed to view this company");
+	}
+
+	try {
 		await prisma.completedTask.deleteMany({
 			where: {
 				companyId: companyId,
 				taskId: taskId,
 			},
 		});
-	} catch (e) {
-		if (e instanceof Prisma.PrismaClientKnownRequestError) {
-			if (e.code == "P2025") {
-				throw new HttpException(
-					404,
-					"No task with this ID found in the specified company"
-				);
-			}
-		} else {
-			throw new HttpException(500, e.message);
-		}
-	}
+	} catch (e) {}
 }
